move colorFunction to Colors.js
